refactor: create callbacks.py
